refactor(common/models): split trie-model implementation into separate components
